src/shodan-producer/requirements.txt
# Pip Requirements File
shodan 
kafka-python
python-dotenv